experiments/sql_test_kit_sales_aggregation/requirements.txt
sql-test-kit
psycopg2-binary
pandas
